requirements.txt
# Core dependencies (maintained in pyproject.toml)
# This file is for backwards compatibility and simple pip installations
# For development, use: pip install -e ".[dev]"

langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
fastapi>=0.115.0
langserve[all]>=0.3.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
google-cloud-storage>=2.18.0
pygithub>=2.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
